backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
pillow==10.1.0
opencv-python==4.8.1.78
mediapipe==0.10.8
scipy==1.11.4
piexif==1.1.3
numpy==1.26.2
